pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "Code_generation_project"
version = "0.1.0"
description = "Fastapi based web application which takes prompt and generate a code basee on it"
authors = ["ramanand <ramanandh.sb@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
transformers = "==4.38.2"
sentencepiece = "==0.2.0"
absl-py = "==2.1.0"
black = "==24.3.0"
ctransformers = "==0.2.27"
fastapi = "==0.110.0"
langchain = "==0.1.13"
langchain-community = "==0.0.29"
numpy = "==1.26.4"
pydantic = "==2.6.4"
uvicorn = "==0.29.0"
uvloop = "==0.19.0"

[tool.poetry.dev-dependencies]

